make code clean and extendable
